requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
gym>=0.26.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
pytest>=7.3.1
tqdm>=4.65.0
scikit-learn>=1.2.0
tensorboard>=2.13.0
jsonlines>=3.1.0
psutil>=5.9.0
gputil>=1.4.0
pytest>=7.3.1
pytest-cov>=4.1.0